Pow works much better, exp and ln are fixed. still can find bugs in pow. Magor fix needed!

+++ src/pow_sqrt/s21_pow_sqrt.h
@@ -18,7 +18,7 @@
 #endif
 
 long double s21_pow(double base, double p);
-long double s21_pow_calculation(double base, double p);
+long double s21_pow_calculation(double base, long double p);
 long double s21_pow_int(long double base, long double p);
 long double s21_pow_double(double base, double p);
 long double s21_sqrt(double base);
